backend/utils/chatbot/requirements.txt
langgraph>=0.0.19
langchain_groq>=0.1.0
langchain_core>=0.1.27
python-dotenv>=1.0.0
